Move getSecretValue and getConfigMapValue to pkg kube
This change moves the methods getSecretValue and getConfigMapValue into a new
package kube. Further logging setup for tests are moved into package test and
controller-runtime is upgrade to latest.
